% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbpR_bills.R, R/cbpR_bills_all.R
\name{cbpR_bills}
\alias{cbpR_bills}
\alias{cbpR_bills}
\title{Request bills based on the query parameters}
\usage{
cbpR_bills(bill_id = NULL, bill_type = NULL, name = NULL,
  key = Sys.getenv("CBPR_API_KEY"))

cbpR_bills(bill_id = NULL, bill_type = NULL, name = NULL,
  key = Sys.getenv("CBPR_API_KEY"))
}
\arguments{
\item{name}{character. The last name of a representative or senator.}

\item{key}{character. A valid API key which can be requested at .}

\item{id}{character. The id of a certain bill.}

\item{offset}{integer.}

\item{limit}{integer.}

\item{key}{character.}

\item{key}{character. A valid API key which can be requested at .}

\item{id}{character. The id of a certain bill.}

\item{name}{character. The last name of a representative or senator.}
}
\description{
Request bills based on the query parameters

Request automatically all bills based on the query parameters
}
