% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbpR_bills_all.R
\name{cbpR_bills_all}
\alias{cbpR_bills_all}
\title{Request all bills based on the query parameters}
\usage{
cbpR_bills_all(q = NULL, order = NULL, key = Sys.getenv("CBPR_API_KEY"))
}
\arguments{
\item{q}{character. The search query. See \url{https://cbapi.dataflood.de/documentation/endpoints/bills.html} for detailed description.}

\item{order}{character.}

\item{key}{character. A valid API key which can be requested at \url{https://cbapi.dataflood.de/start}.}
}
\value{
The output is a cbpR API object. Which includes meta information about the query and the respective data set that is defined in the parameter \code{q}.
}
\description{
\code{cbpR_bills} returns all US Congressional Bills depending on the provided query regardless of the number of rows. The function uses the API of the extensive Congressional Bills Project. Information about the Congressional Bills Project can be found on the official website: \url{http://congressionalbills.org}.
}
