% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/init.R
\name{checkCbp}
\alias{checkCbp}
\title{Check if the data directories are set}
\usage{
checkCbp()
}
\value{
\code{true} if the function concludes without error
}
\description{
This function produces an error if the data directories were not set.
}
\examples{
\dontrun{checkCbp()}
}

