% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/download.R
\name{downloadCbp}
\alias{downloadCbp}
\title{Download the data}
\usage{
downloadCbp()
}
\value{
\code{true} if the function concludes without error
}
\description{
This function downloads the data from the Census website. All
Downloads are stored in the directory that was specified using
\code{setCbpPath()}
}

