% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/init.R
\name{getCbpPath}
\alias{getCbpPath}
\title{Get the path to the data directories}
\usage{
getCbpPath()
}
\value{
a list with the paths to the data directories
}
\description{
Get the path to the data directories
}

