% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/population.R
\name{getPopulationData}
\alias{getPopulationData}
\title{Returns a long panel of the population data}
\usage{
getPopulationData()
}
\value{
a data frame with population data for micro- and metropolitan
statistical areas
}
\description{
Reshapes the population data for the Core-Based Statistical Areas into
a long panel.
}
\examples{
\dontrun{getPopulationData()}
}

