% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/states.R
\name{getStatesFips}
\alias{getStatesFips}
\title{State abbreviations, fips codes and state names}
\usage{
getStatesFips()
}
\value{
a data frame with state abbreviations, fips codes, and state names
}
\description{
State abbreviations, fips codes and state names
}

