% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/init.R
\name{setCbpPath}
\alias{setCbpPath}
\title{Set the environmental variables to the data directories}
\usage{
setCbpPath(data_in, data_out)
}
\arguments{
\item{data_in}{a string with the path to the directory where the downloaded
data will be stored}

\item{data_out}{a string with the path to the directory where the processed
data will be stored}
}
\value{
\code{true} if the function concludes successfully
}
\description{
This function sets the environmental variables that point to the directories
where the downloaded and processed data is stored
}

