% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbpR_bills.R
\name{cbpR_bills}
\alias{cbpR_bills}
\title{Request bills based on the query parameters}
\usage{
cbpR_bills(key, id = NULL, name = NULL)
}
\arguments{
\item{key}{character. A valid API key which can be requested at .}

\item{id}{character. The id of a certain bill.}

\item{name}{character. The last name of a representative or senator.}
}
\description{
Request bills based on the query parameters
}
