% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbpR_bills_all.R
\name{cbpR_bills_all}
\alias{cbpR_bills_all}
\title{Get all query parameters.}
\usage{
cbpR_bills_all(bill_id = NULL, bill_type = NULL, name = NULL,
  key = Sys.getenv("CBPR_API_KEY"))
}
\arguments{
\item{bill_id}{character. The id of a certain bill.}

\item{bill_type}{character. Bill type ('hr' (house bill); 's' (senate bill); 'hres' (House resolution); 'sres' (Senate resolution); 'hcon' (House Concurrent Resolution); 'scon' (Senate Concurrent Resolution); 'hjres' (House Joint Resolution); 'sjres' (Senate Joint Resolution).}

\item{name}{character. The last name of a representative or senator.}

\item{key}{character. A valid API key which can be requested at URL.}
}
\description{
Get all query parameters.
}
