% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/download.R
\name{downloadFile}
\alias{downloadFile}
\title{Downloads a file and checks the size}
\usage{
downloadFile(url, filename, size)
}
\arguments{
\item{url}{}

\item{filename}{to be used locally}

\item{size}{of the file in bytes}
}
\value{
\code{true} if the function concludes without error
}
\description{
This function downloads the specified file and verifies its size.
Since the Census servers don't always work without error, the function tries
downloading each file ten times before it gives up.
}

